AAAAAAAAA

+++ migrations/2024-02-05-233306_users/up.sql
@@ -1,5 +1,5 @@
 -- Your SQL goes here
 CREATE TABLE IF NOT EXISTS users (
-  uid SERIAL PRIMARY KEY NOT NULL,
+  uid BIGINT PRIMARY KEY NOT NULL,
   groups INTEGER[]
 );
